+++ postWRF/__init__.py
@@ -1,0 +1,1 @@
+from postWRF.main import WRFEnviron
