+++ postWRF/postWRF/figure.py
@@ -12,7 +12,7 @@ import pdb
 import os
 
 # Custom imports
-import WEM.utils.utils as utils
+import WEM.utils as utils
 
 class Figure:
     def __init__(self,config,wrfout):
