+++ postWRF/postWRF/metconstants.py
@@ -1,0 +1,10 @@
+Tb = 300.0 # Base temperature
+Tz = 273.15 # 0 C in Kelvin
+L = 2.501e6 # Latent heat of vaporisation
+R = 287.04  # gas constant air  
+Rv = 461.5  # gas constant vapour
+cp = 1005.0
+cv = 718.0
+kappa = (cp-cv)/cp
+g = 9.81
+P0 = 10**5 # Reference pressure
